requirements.txt
dl-translate
tqdm
beautifulsoup4
torch
nltk
PyYAML